pyproject.toml
[project]
name = "deep-sprite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absl-py==2.1.0",
    "aiohappyeyeballs==2.4.3",
    "aiohttp==3.10.10",
    "aiosignal==1.3.1",
    "antlr4-python3-runtime>=4.9.3",
    "argcomplete==3.5.1",
    "asgiref>=3.11.1",
    "asttokens>=3.0.1",
    "astunparse==1.6.3",
    "async-timeout==4.0.3",
    "attrs==24.2.0",
    "boto==2.49.0",
    "bottleneck>=1.6.0",
    "brotli>=1.2.0",
    "cachetools==5.5.0",
    "certifi>=2026.2.25",
    "cffi==1.17.1",
    "charset-normalizer>=3.4.6",
    "comm>=0.2.3",
    "contourpy>=1.3.3",
    "crcmod==1.7",
    "cryptography==43.0.1",
    "cycler>=0.12.1",
    "debugpy>=1.8.20",
    "decorator>=5.2.1",
    "django>=6.0.3",
    "django-environ>=0.13.0",
    "exceptiongroup>=1.3.1",
    "executing>=2.2.1",
    "fastcore==1.7.19",
    "fasteners==0.19",
    "filelock>=3.25.2",
    "flatbuffers==24.3.25",
    "fonttools>=4.62.1",
    "frozenlist==1.4.1",
    "fsspec==2024.10.0",
    "gast==0.6.0",
    "gcs-oauth2-boto-plugin==3.2",
    "gdown>=5.2.1",
    "gmpy2>=2.3.0",
    "google-apitools==0.5.32",
    "google-auth==2.17.0",
    "google-auth-httplib2==0.2.0",
    "google-pasta==0.2.0",
    "google-reauth==0.1.1",
    "grpcio==1.66.2",
    "gsutil==5.31",
    "h5py>=3.16.0",
    "hf-xet==1.1.3",
    "httplib2==0.20.4",
    "huggingface-hub==0.32.5",
    "hydra-core>=1.3.2",
    "idna>=3.11",
    "importlib-metadata>=8.7.1",
    "importlib-resources>=6.5.2",
    "ipykernel>=7.2.0",
    "ipython>=9.11.0",
    "jedi>=0.19.2",
    "jinja2>=3.1.6",
    "joblib>=1.5.3",
    "jupyter-client>=8.8.0",
    "jupyter-core>=5.9.1",
    "keras==3.6.0",
    "kiwisolver>=1.5.0",
    "kornia>=0.8.2",
    "libclang==18.1.1",
    "lovely-numpy==0.2.13",
    "lovely-tensors==0.1.17",
    "markdown==3.7",
    "markdown-it-py==3.0.0",
    "markupsafe>=3.0.3",
    "matplotlib==3.9.2",
    "matplotlib-inline>=0.2.1",
    "mdurl==0.1.2",
    "mkl-fft>=1.3.14",
    "mkl-random>=1.2.10",
    "mkl-service>=2.4.1",
    "ml-dtypes==0.4.1",
    "monotonic==1.6",
    "mpmath>=1.3.0",
    "multidict==6.1.0",
    "namex==0.0.8",
    "nest-asyncio>=1.6.0",
    "networkx>=3.6.1",
    "numexpr>=2.14.1",
    "numpy>=1.26.4",
    "oauth2client==4.1.3",
    "omegaconf>=2.3.0",
    "opencv-python==4.10.0.84",
    "opt-einsum==3.4.0",
    "optree==0.13.0",
    "packaging>=26.0",
    "pandas>=2.3.3",
    "parso>=0.8.6",
    "pexpect>=4.9.0",
    "pickleshare>=0.7.5",
    "pillow>=12.1.1",
    "platformdirs>=4.9.4",
    "ply==3.11",
    "prompt-toolkit>=3.0.52",
    "propcache==0.2.0",
    "protobuf==4.25.5",
    "psutil>=7.2.2",
    "ptyprocess>=0.7.0",
    "pure-eval>=0.2.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pycparser==2.22",
    "pygments>=2.19.2",
    "pyopenssl==24.2.1",
    "pyparsing>=3.3.2",
    "pyqt5==5.15.10",
    "pyqt5-sip>=12.18.0",
    "pysocks>=1.7.1",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2026.1.post1",
    "pyu2f==0.1.5",
    "pyyaml>=6.0.3",
    "pyzmq>=27.1.0",
    "regex==2024.11.6",
    "requests>=2.32.5",
    "retry-decorator==1.1.1",
    "rich==13.9.2",
    "rsa==4.7.2",
    "safetensors==0.5.3",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "sip>=6.15.2",
    "six>=1.17.0",
    "sqlparse>=0.5.5",
    "stack-data>=0.6.3",
    "sympy>=1.13.1",
    "tensorboard==2.17.1",
    "tensorboard-data-server==0.7.2",
    "tensorflow==2.17.0",
    "tensorflow-io-gcs-filesystem==0.37.1",
    "termcolor==2.5.0",
    "threadpoolctl>=3.6.0",
    "tokenizers==0.21.1",
    "tomli>=2.4.0",
    "torch==2.5.1",
    "torchaudio==2.5.1",
    "torchvision==0.20.1",
    "tornado>=6.5.5",
    "tqdm==4.67.1",
    "traitlets>=5.14.3",
    "transformers==4.52.4",
    "triton==3.1.0",
    "typing-extensions>=4.15.0",
    "tzdata>=2025.3",
    "unicodedata2>=17.0.1",
    "urllib3>=2.6.3",
    "wcwidth>=0.6.0",
    "werkzeug==3.0.4",
    "wrapt==1.16.0",
    "yarl==1.14.0",
    "zipp==3.20.2",
]
